+++ babel.config.js
@@ -20,7 +20,7 @@ module.exports = {
 			"path": ".env",
 			"blacklist": null,
 			"whitelist": null,
-			"safe": false,
+			"safe": false, 
 			"allowUndefined": true
 		}]
 	],
