+++ ios/ParkingFinder/AppDelegate.h
@@ -1,6 +1,8 @@
 #import <RCTAppDelegate.h>
 #import <UIKit/UIKit.h>
+#import <YandexMapsMobile/YMKMapKitFactory.h>
 
 @interface AppDelegate : RCTAppDelegate
+
 
 @end
