+++ ios/ParkingFinder/AppDelegate.mm
@@ -1,7 +1,7 @@
 #import "AppDelegate.h"
 #import <Firebase.h>
-
 #import <React/RCTBundleURLProvider.h>
+#import <YandexMapsMobile/YMKMapKitFactory.h>
 
 @implementation AppDelegate
 
@@ -11,12 +11,22 @@
    if ([FIRApp defaultApp] == nil) {
         [FIRApp configure];
       }
+
+  self.initialProps = @{};
+
+
+    [YMKMapKit setLocale:@"ru_RU"];
+    [YMKMapKit setApiKey:@"API_KEY"];
+
+
   // You can add your custom initial props in the dictionary below.
   // They will be passed down to the ViewController used by React Native.
-  self.initialProps = @{};
+
 
   return [super application:application didFinishLaunchingWithOptions:launchOptions];
 }
+
+
 
 - (NSURL *)sourceURLForBridge:(RCTBridge *)bridge
 {
