+++ package.json
@@ -26,8 +26,8 @@
     "react-native-reanimated": "^3.14.0",
     "react-native-safe-area-context": "^4.10.8",
     "react-native-screens": "^3.32.0",
-    "react-native-yamap": "^4.6.0",
     "react-native-svg": "^15.4.0",
+    "react-native-yamap": "^4.6.0",
     "zustand": "^4.5.4"
   },
   "devDependencies": {
