+++ src/App/navigation/navigators/BottomTabNavigator.tsx
@@ -23,3 +23,4 @@ const BottomTabNavigator = () => {
 };
 
 export default BottomTabNavigator;
+
