+++ src/screens/AddParkScreen/ui/AddParkScreen.tsx
@@ -1,6 +1,5 @@
 import { View, Text } from "react-native";
 import React from "react";
-import {styles} from "./AddParkScreen.styles";
 
 export const AddParkScreen = () => {
 	return (
