--- src/screens/AuthScreen/index.ts
@@ -1,3 +1,0 @@
-import AuthScreen from "./ui/AuthScreen";
-
-export {AuthScreen};
