+++ src/screens/OTPVerification/index.ts
@@ -1,0 +1,1 @@
+export {OTPVerification} from "./ui/OTPVerification";
