--- src/shared/constants/appColors.ts
@@ -1,12 +1,0 @@
-export const appColors = {
-    primary: 'rgb(40, 167, 69)',
-    secondary: 'rgb(0, 123, 255)',
-    background: 'rgb(255, 255, 255)',
-    backgroundLight: 'rgb(240, 240, 240)',
-    border: 'rgb(108, 117, 125)',
-    text: 'rgb(33, 37, 41)',
-    textLight: 'rgb(108, 117, 125)',
-    danger: 'rgb(220, 53, 69)',
-    disabled: 'rgb(108, 117, 125)',
-    mutedGreen: 'rgb(134, 188, 145)',
-};
