+++ src/shared/ui/Components/CustomCheckBox/ui/CustomCheckBox.style.ts
@@ -1,0 +1,13 @@
+import { StyleSheet } from "react-native";
+import { StyleGuide } from "@shared/ui";
+
+export const styles = StyleSheet.create({
+	checkbox: {
+		backgroundColor: StyleGuide.WHITE,
+		width: 30,
+		height: 30,
+		borderRadius: 5,
+		borderColor: StyleGuide.GREY,
+		borderWidth: 1
+	}
+});
