+++ src/shared/ui/Components/CustomImage/ui/CustomImage.styles.ts
@@ -1,8 +1,8 @@
 import { StyleSheet } from "react-native";
 
 export const styles = StyleSheet.create({
-    checkedImg: {
-        width: 27,
-        height: 27
-    }
+	checkedImg: {
+		width: 27,
+		height: 27
+	}
 });
