--- src/shared/ui/CustomTextInput/ui/CustomTextInput.style.ts
@@ -1,12 +1,0 @@
-import { StyleSheet } from "react-native";
-
-export const styles = StyleSheet.create({
-    inputBox: {
-        width: '100%',
-        justifyContent: 'center',
-        paddingHorizontal: 10
-    },
-    input: {
-        width: '100%',
-    }
-});
