+++ src/shared/ui/assets/images/checkMark.png
