+++ src/shared/ui/stylesConsts/stylesConsts.ts
@@ -4,8 +4,7 @@ export enum StyleGuide {
     GREY = "#e3e3e3",
     WHITE ="#FFFFFF",
     BLACK= "#130F26",
-    GREEN ="#33AD5F",
-
+    PRIMARY ="#33AD5F",
 
     //TEXT COLORS
     TEXT_GREY = "#7D7D91",
