+++ src/widgets/SignUp/ui/SignUp.tsx
@@ -4,12 +4,7 @@ import { useUserStore } from "@entities/user";
 import { CustomInput, StyleGuide } from "@shared/ui";
 import { CustomButton } from "@shared/ui/Components/CustomButton/ui/CustomButton.tsx";
 
-interface SignUpProps {
-}
-
-
-export const SignUp = (props: SignUpProps) => {
-	const {} = props;
+export const SignUp = () => {
 	const [email, setEmail] = useState("");
 	const [password, setPassword] = useState("");
 	const signUp = useUserStore(state => state.signUp);
@@ -21,7 +16,7 @@ export const SignUp = (props: SignUpProps) => {
 		<KeyboardAvoidingView>
 			<CustomInput title={"Email"} value={email} setValue={setEmail}/>
 			<CustomInput title={"Password"} value={password} setValue={setPassword} isPassword/>
-			<CustomButton title={"Sign Up"} onPress={handleSignUp} color={StyleGuide.GREEN}/>
+			<CustomButton title={"Sign Up"} onPress={handleSignUp} color={StyleGuide.PRIMARY}/>
 		</KeyboardAvoidingView>
 	);
 };
